pyproject.toml
[tool.poetry]
name = "langchain-exploration"
version = "0.1.0"
description = "Dependencies for langchain-exploration"
authors = ["Ritwick Manatkar <ritwickmanatkar@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
streamlit = "^1.41.1"
langchain = "^0.3.14"
openai = "^1.59.8"
langchain-community = "^0.3.14"
langchain-openai = "^0.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
